Implement SLP-29 for-in loops and SLP-30 iterator protocol
This commit adds `for ... in` loop syntax for sequence iteration with support for value, reference, and pointer captures, optional key binding, and value discard. Includes comprehensive documentation, parser support, type checking, code generation, and formatting. Also adds the `in` keyword to the lexer and defines three new diagnostics for invalid source expressions, invalid key references, and invalid value bindings.

+++ src/typecheck/api.c
@@ -22,7 +22,7 @@ int SLConstEvalSessionInit(
     SLConstEvalSession** outSession,
     SLDiag* _Nullable diag) {
     SLConstEvalSession* session;
-    uint16_t*           savedFlags = NULL;
+    uint32_t*           savedFlags = NULL;
     uint32_t            i;
     if (outSession == NULL) {
         return -1;
@@ -39,8 +39,8 @@ int SLConstEvalSessionInit(
         return -1;
     }
     if (ast->len > 0) {
-        savedFlags = (uint16_t*)SLArenaAlloc(
-            arena, (uint32_t)sizeof(uint16_t) * ast->len, (uint32_t)_Alignof(uint16_t));
+        savedFlags = (uint32_t*)SLArenaAlloc(
+            arena, (uint32_t)sizeof(uint32_t) * ast->len, (uint32_t)_Alignof(uint32_t));
         if (savedFlags == NULL) {
             SLTCSetDiag(diag, SLDiag_ARENA_OOM, 0, 0);
             return -1;
